backend/api/requirements.txt
aniso8601==7.0.0
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT==0.3.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector==2.2.9
numpy==1.17.0
PyJWT==1.4.2
PyMySQL==0.9.3
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.6
Werkzeug==0.15.5
